LayoutTests/fast/js/dfg-cse-cfa-discrepancy-expected.txt

Tests stability when the CFA finds a constant in block-local data flow over captured variables but the CSE doesn't because of differences in heap modeling.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS foo(i, i + 1) is 162
PASS successfullyParsed is true

TEST COMPLETE
